Added to documentation comments, changed stages() argument order.
Documented the fact that the pipeline tasks may be asynchronous, through
the use of promises.

Several methods implicitly define pipelines and stages if they are not
already defined. Documented this behavior since it's intentional.

In the module overview documentation, the stages() method arguments are
described as (pipeline-name, stage-name-array), but the actual method
definition is (stage-name-array, pipeline-name). All the other methods
place the pipeline-name as the first argument, so I changed the stages()
definition to match. Of course, this is a breaking change if there is
any code actually using this method.

Didn't document the fulfilled value of the promise returned by the run()
method. (Because it's not clear this value is useful, and it's hard to
describe.)